pyproject.toml
[project]
name = "cag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-google-genai>=2.0.10",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.2",
]
